Add the base implementation for the new repository
The `Repository` class is the new interface to interact with the file
repository. It contains a file hierarchy in memory to now what is
contained within it. A new instance can be constructed from an existing
hierarchy using the `from_serialized` class method.

The actual storing of the objects in a repository is delegated to a
backend repository instance, represented by `AbstractRepositoryBackend`.
This backend class is only tasked with consuming byte streams and
storing them as objects in some repository under a unique key that
allows the object to be retrieved at a later time. The backend
repository is not expected to keep a mapping of what it contains. It
should merely provide the interface to store byte streams and return
their content, given their unique corresponding key.

This also means that the internal virtual hierarchy of a ``Repository``
instance does not necessarily represent all the files that are stored by
repository backend. The repository exposes a mere subset of all the file
objects stored in the backend. This is why object deletion is also
implemented as a soft delete, by default, where the files are just
removed from the internal virtual hierarchy, but not in the actual
backend. This is because those objects can be referenced by other
instances.

The interface is implemented for the `disk-objectstore` which is an
efficient key-value store on the local file system. This commit also
provides a sandbox implementation which represents a temporary folder on
the local file system.